Move RandomModuleGenerator to separate file
